Refactor accordion component
